code/ai-connect/pyproject.toml
[project]
name = "ai-connect"
version = "0.1.0"
description = "Websocket connection to the AI backend"
authors = [
    {name = "sachink108",email = "sachink108@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.68.2,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "numpy (>=2.2.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
